US7 Complete (#80)
* Remove Driver placeholder data

The placeholders for Drivers have been replaced with the DriverColl and Driver classes

* Create Add Driver JDialog

The JDialog to add a driver to the system has been created

* Remove unused code from NewDriverDialog

Removed unused imports that were used for the original class NewDriverDialog was derived from

* Edit JDialog

US50 Task 81/101
Made a slight edit to the JDialog for adding a new driver to remove the empty space in the dialog

* Implement Edit and Delete Driver

US50 Task99/100
User can now edit or delete a driver by selecting a driver from the table and right-clicking to open a popup menu

* General code improvements

US50 - No Task
Cleaned up some code, removed some unused code, conformed code to QualityPolicy

* Rename branch to US80

Renamed this branch to reflect the change in the User Story Number of Taiga

* Create schedule popup

US80 Task102
Created the schedule popup menu

* Add DriverColl and TourColl to CurrentProject

US80 All Tasks
DriverColl and TourColl objects were added to the CurrentProject class so that all project classes have access to the same instance of these classes

* Remove duplicate methods from Driver

US80 All Tasks
I had previously added some methods to Driver that were needed to fully implement the DriverTable. Brian's latest commits also included these methods, so I have removed the duplicates

* Update DriverPanel

US80 All Tasks
Updated the DriverPanel and all it's child Components to work with the updated DriverColl

* Fix DriverDialog

US50 All Tasks
After refactoring to work with the latest changes in Dev, DriverDialog no longer worked as intended. This commit fixes the issue

* Implement edit driver schedule

US80 Task102
A user can now add or remove tours from a driver's schedule

* Display schedule when driver is selected

US80 Task 98
When a driver is selected from the DriverTable, their schedule will be displayed on the ScheduleTable
User Story is now complete and ready for testing

* Edit code comments and remove unused imports

US80
Changed some code comments to reflect recent changes to code, and removed usnused import statements from some classes

* Complete US80

All Tasks
This user story is now complete and ready for review

* Correct errors that caused to TravisCI to fail

* Fix more errors that caused Travis CI to fail

* US80 test: update Driver/Tour

* Revert several UT changes

* Reverted some changes

* US80 Completed (For real this time)

Worked with Brian to get things working correctly on both ends, made a slight improvement to the DriverTourDialog, and resolved a new issue that occured when deleting a driver

* Create Bus UI components (US7: Task 9 & 10)

Driver UI Components were copied and ranmed to be ralted to Buses; began converting BusPanel from displaying Driver data to Bus data

* Remove all Driver references (US7: Tasks 9&10)

Since the Bus UI componenets were copied from the Driver components, they all containted references to drivers and used driver objects. These have all been changed to Buses and Bus objects

* Finish conversion of BusDialog (US7: Tasks 9&10)

Finished the conversion of the BusDialog class from the DriverDialog class

* Remove unused code block and fix a incorrect Javadoc (US80: No Task)

Noticed I didn't have param tags on some methods and removed a piece of code that did nothing

* Complete Task 9 (US7)

A lost of Buses is now properly displayed in the Buses tab

* Complete Tasks 103 & 104 (US7)

Users can now create and delete a Bus

* US7 - Update classes based on changes to Bus.java

* US80 - Task 102: Update Dialog for scheduling a tour

Due to changes to how a Tour is added to a Bus, this changed was required to properly load the list of tours that could be shceduled

* US7 - No Task: Update comments

* US7: Update Bus/Tour association on load

Co-authored-by: Brian Pape <bpape1@asu.edu>